add test config
