+++ restdemo/__init__.py
@@ -10,14 +10,14 @@ from restdemo.model.user import User as UserModel
 from restdemo.model.tweet import Tweet
 from restdemo.resource.user import User, UserList
 from restdemo.resource.tweet import Tweet
-from restdemo.config import Config
+from restdemo.config import app_config
 
 jwt = JWT(None, UserModel.authenticate, UserModel.identity)
 
-def create_app():
+def create_app(config_name='development'):
     app = Flask(__name__)
     api = Api(app)
-    app.config.from_object(Config)
+    app.config.from_object(app_config[config_name])
     db.init_app(app)
     migrate = Migrate(app, db)
     jwt.init_app(app)
