+++ restdemo/model/base.py
@@ -1,0 +1,21 @@
+from restdemo import db
+
+
+class Base(db.Model):
+
+    __abstract__ = True
+
+    def as_dict(self):
+        """使序列化返回字典"""
+        return {c.name: getattr(self, c.name) for c in self.__table__.columns}
+
+    def add(self):
+        db.session.add(self)
+        db.session.commit()    
+
+    def delete(self):
+        db.session.delete(self)
+        db.session.commit()
+
+    def update(self):
+        db.session.commit()
