+++ restdemo/model/tweet.py
@@ -1,8 +1,9 @@
 from sqlalchemy import ForeignKey, func
 
 from restdemo import db
+from restdemo.model.base import Base
 
-class Tweet(db.Model):
+class Tweet(Base):
 
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, ForeignKey('user.id'))
@@ -14,11 +15,8 @@ class Tweet(db.Model):
             self.user_id, self.body
         )
 
-    def add(self):
-        db.session.add(self)
-        db.session.commit()
-
     def as_dict(self):
+        """override as_dict"""
         t = {c.name: getattr(self, c.name) for c in self.__table__.columns}
         t['created_at'] = t['created_at'].isoformat()
         return t
