+++ restdemo/model/user.py
@@ -6,9 +6,10 @@ from werkzeug.security import generate_password_hash, check_password_hash
 from sqlalchemy.orm import relationship
 
 from restdemo import db
+from restdemo.model.base import Base
 
 
-class User(db.Model):
+class User(Base):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(64), unique=True)
     password_hash = db.Column(db.String(128))
@@ -21,26 +22,11 @@ class User(db.Model):
             self.id, self.username
         )
 
-    def as_dict(self):
-        """使序列化返回字典"""
-        return {c.name: getattr(self, c.name) for c in self.__table__.columns}
-
     def set_password(self, password):
         self.password_hash = generate_password_hash(password)
 
     def check_password(self, password):
         return check_password_hash(self.password_hash, password)
-
-    def add(self):
-        db.session.add(self)
-        db.session.commit()    
-
-    def delete(self):
-        db.session.delete(self)
-        db.session.commit()
-
-    def update(self):
-        db.session.commit()
 
     @staticmethod
     def get_by_username(username):
